removed nonessential tests and changed blinky for f429zi
