--- test/tests/ADC_DMA.c
@@ -1,13 +1,0 @@
-#include "stm32xx_hal.h"
-
-#include <stdio.h>
-#include <string.h>
-
-int main(void) {
-    HAL_Init();
-    
-    SystemClock_Config();
-    MX_ADC1_Init();
-
-    HAL_ADC_DMA();
-}
