+++ test/tests/blinky.c
@@ -1,21 +1,38 @@
 #include "stm32xx_hal.h"
 
+static void MX_GPIO_Init(void);
+
 int main(){
     HAL_Init();
+    SystemClock_Config();
 
-    GPIO_InitTypeDef led_config = {
-        .Mode = GPIO_MODE_OUTPUT_PP,
-        .Pull = GPIO_NOPULL,
-        .Pin = GPIO_PIN_5
-    };
-    
-    __HAL_RCC_GPIOA_CLK_ENABLE(); // enable clock for GPIOA
-    HAL_GPIO_Init(GPIOA, &led_config); // initialize GPIOA with led_config
+    MX_GPIO_Init();
 
     while(1){
-        HAL_GPIO_TogglePin(GPIOA, GPIO_PIN_5);
+        HAL_GPIO_TogglePin(GPIOA, GPIO_PIN_3);
+        //HAL_GPIO_TogglePin(GPIOA, GPIO_PIN_5);
         HAL_Delay(1000);
     }
 
     return 0;
+}
+
+static void MX_GPIO_Init(void)
+{
+  GPIO_InitTypeDef GPIO_InitStruct = {0};
+
+  /* GPIO Ports Clock Enable */
+  __HAL_RCC_GPIOA_CLK_ENABLE();
+
+  /*Configure GPIO pin Output Level */
+
+  /*Configure GPIO pin : PA5 */
+  GPIO_InitStruct.Pin = GPIO_PIN_3;
+  GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
+  GPIO_InitStruct.Pull = GPIO_NOPULL;
+  GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
+  HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+/* USER CODE BEGIN MX_GPIO_Init_2 */
+/* USER CODE END MX_GPIO_Init_2 */
 }
